+++ changelog.md
@@ -5,6 +5,7 @@
 
 ## 2026-03
 * [changed] Better smoke test with factory
+* [added] Identity code can be defined by regex in the work group settings
 
 ## 2026-02
 * [added] Dark mode